Fusion: fixing mismatch in default parameter to nview
git-svn-id: http://svn.boost.org/svn/boost/trunk@56604 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/fusion/view/nview/detail/nview_impl.hpp
@@ -29,7 +29,7 @@
 namespace boost { namespace fusion { namespace result_of
 {
     template <typename Sequence
-      , BOOST_PP_ENUM_PARAMS_WITH_A_DEFAULT(FUSION_MAX_VECTOR_SIZE, int I, INT_MAX)>
+      , BOOST_PP_ENUM_PARAMS_WITH_A_DEFAULT(FUSION_MAX_VECTOR_SIZE, int I, LONG_MAX)>
     struct as_nview
     {
         typedef mpl::vector_c<
